- 이제 안쓰는 임시 api 지움 - 사용자 정의 exception 생성 	- IdNotFoundException : db에 없는 데이터를 호출할 때 	- IndexOutOfRangeException : db에 존재하는 데이터 범위를 넘어갈때 - api 응답 공통 포맷 작성 	- 정상/에러시 동일한 형태로 응답 처리 	- 정상/에러 구분 용이 - DTO 클래스 변수 타입 Date -> LocalDate

+++ src/main/java/com/officehub/officehubserver/OfficeHubServer/service/NoticeServiceImpl.java
@@ -1,9 +1,10 @@
 package com.officehub.officehubserver.OfficeHubServer.service;
 
 import com.officehub.officehubserver.OfficeHubServer.dto.NoticeDto;
+import com.officehub.officehubserver.OfficeHubServer.exception.IdNotFoundException;
+import com.officehub.officehubserver.OfficeHubServer.exception.IndexOutOfRangeException;
 import com.officehub.officehubserver.OfficeHubServer.repository.NoticeMapper;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.HashMap;
 import java.util.List;
@@ -18,31 +19,59 @@ public class NoticeServiceImpl implements NoticeService{
         this.noticeMapper = noticeMapper;
     }
 
-    public List<NoticeDto> getNoticeList(int offset, int size) {
+    public List<NoticeDto> getNoticeList(int offset, int size) throws IndexOutOfRangeException{
+        int totalCountOfNotices = noticeMapper.getCountOfNotices();  // 공지사항 총 개수
+
+        // 출력하고자하는 결과가 db의 데이터들의 범위를 넘어갈시 에러 발생
+        if (totalCountOfNotices <= offset) {
+            throw new IndexOutOfRangeException("index is out of range");
+        }
+
         Map<String, Integer> map = new HashMap<>();
         map.put("offset", offset);
         map.put("size", size);
         return noticeMapper.getNoticeList(map);
     }
 
-    public NoticeDto getNoticeById(int noticeId) {
-        return noticeMapper.getNoticeById(noticeId);
+    public NoticeDto getNoticeById(int noticeId) throws IdNotFoundException{
+        NoticeDto noticeDto = noticeMapper.getNoticeById(noticeId);
+
+        // 파라미터로 받은 id의 데이터가 없을시 에러 발생
+            if (noticeDto == null) {
+                throw new IdNotFoundException("current id is not found");
+        }
+
+        return noticeDto;
     }
 
     @Override
     public void insertNotice(NoticeDto dto) {
+
         noticeMapper.insertNotice(dto);
     }
 
     @Override
-    public void updateNotice(NoticeDto dto) {
+    public void updateNotice(NoticeDto dto) throws IdNotFoundException{
+        NoticeDto noticeDto = noticeMapper.getNoticeById(dto.getNoticeId());
+
+        // 파라미터로 받은 id의 데이터가 없을시 에러 발생
+        if (noticeDto == null) {
+            throw new IdNotFoundException("current id is not found");
+        }
+
         noticeMapper.updateNotice(dto);
     }
 
     @Override
     public void deleteNotice(int noticeId) {
+        NoticeDto noticeDto = noticeMapper.getNoticeById(noticeId);
+
+        // 파라미터로 받은 id의 데이터가 없을시 에러 발생
+        if (noticeDto == null) {
+            throw new IdNotFoundException("current id is not found");
+        }
+
         noticeMapper.deleteNotice(noticeId);
     }
-
 
 }
